Support Newer Versions of the redis Gem (#1)
### What This Fixes ###
Newer versions of the redis gem include bug fixes, and will likely include clustering support in the near future. (4.1.0.beta1 includes this support.) The cb2 gem uses get, set, del, zremrangebyscore, zadd, zcard and flushdb methods. As these methods have not changed in newer versions of the redis gem, the only change necessary is to update the cb2 gemspec.

This change also adds a require for 'securerandom' to rolling_window.rb. Without it, the reference to SecureRandom is undefined, at least on Ruby 2.3.7.

Finally, added a development dependency for minitest to the gemspec. It appears that this library was included in older versions of Ruby, but is now provided as a separate gem.

### Testing ###
* Ran unit tests.
* We are using this in production (recent version of redis-rb master with cb2).